requirements.txt
django==3.1
djangorestframework==3.12
flake8==3.8.4
requests==2.25